[docs][dispatcher] Updated mxio_port_bind to use uint64_t
The mxio_port_bind syscall used to use a "uintptr_t" type, but has been
changed to always be 64-bits. Updated documentation to reflect this
fact, and updated dispatcher to avoid casting the key argument to
uintptr_t. The updates to the dispatcher are not necessary, but they
make it clear that the type of 'key' is NOT the size of a pointer.

+++ system/ulib/mxio/dispatcher-v2.c
@@ -59,11 +59,11 @@ static void destroy_handler(mxio_dispatcher_t* md, handler_t* handler) {
 
 static void disconnect_handler(mxio_dispatcher_t* md, handler_t* handler) {
     // unbind, so we get no further messages
-    mx_io_port_bind(md->ioport, (uintptr_t) handler, handler->h, 0);
+    mx_io_port_bind(md->ioport, (uint64_t) handler, handler->h, 0);
 
     // send a synthetic message so we know when it's safe to destroy
     mx_io_packet_t packet;
-    packet.hdr.key = (uintptr_t) handler;
+    packet.hdr.key = (uint64_t) handler;
     packet.signals = MX_SIGNAL_SIGNALED;
     mx_io_port_queue(md->ioport, &packet, sizeof(packet));
 
@@ -82,7 +82,7 @@ again:
             printf("dispatcher: ioport wait failed %d\n", r);
             break;
         }
-        handler_t* handler = (void*) (uintptr_t)packet.hdr.key;
+        handler_t* handler = (void*) (uint64_t)packet.hdr.key;
         if (handler->flags & FLAG_DISCONNECTED) {
             // handler is awaiting gc
             // ignore events for it until we get the synthetic "destroy" event
@@ -174,7 +174,7 @@ mx_status_t mxio_dispatcher_add(mxio_dispatcher_t* md, mx_handle_t h, void* cb, 
 
     mxr_mutex_lock(&md->lock);
     list_add_tail(&md->list, &handler->node);
-    if ((r = mx_io_port_bind(md->ioport, (uintptr_t) handler, h,
+    if ((r = mx_io_port_bind(md->ioport, (uint64_t) handler, h,
                              MX_SIGNAL_READABLE | MX_SIGNAL_PEER_CLOSED)) < 0) {
         list_delete(&handler->node);
     }
